pyproject.toml
[tool.poetry]
name = "nonebot-plugin-genshin-cos"
version = "0.1.6"
description = "米游社原神cos图获取"
authors = ["Cvandia <106718176+Cvandia@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nonebot_plugin_genshin_cos"}]

[tool.poetry.dependencies]
python = ">=3.8,<4"
nonebot2 = ">=2.0.0-beta.1"
nonebot-adapter-onebot = ">=2.0.0-beta.1"
httpx = ">=0.19.0"
nonebot-plugin-apscheduler= ">=0.1.3"
aiofiles = ">=0.7.0"
pydantic = ">=1.10.2"
pathlib = ">=1.0.1"
ujson = ">=5.5.0"
playwright = ">=1.30.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
